requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic[email]
python-jose[cryptography]
passlib
python-multipart
python-dotenv
llmsays
psycopg2-binary
vosk
